docs(core): Encapsulated Blog components Rewrote css to be tailwind specific Removed ImageTheme Component
