+++ nx-dev/ui-blog/src/index.ts
@@ -1,3 +1,6 @@
 export { BlogAuthors, BlogAuthorsProps } from './lib/authors';
 export { BlogEntry, BlogEntryProps } from './lib/blog-entry';
-export { MoreBlogs, MoreBlogsProps } from './lib/more-blogs/more-blogs';
+export { MoreBlogs, MoreBlogsProps } from './lib/more-blogs';
+export { BlogDetails, BlogDetailsProps } from './lib/blog-details';
+export { BlogContainer } from './lib/blog-container';
+export { FeaturedBlogs } from './lib/featured-blogs';
