--- nx-dev/ui-blog/src/lib/more-blogs/more-blogs.module.css
@@ -1,52 +1,0 @@
-.layout {
-  display: grid;
-  column-gap: 24px;
-  grid-template-areas: 'a a a a a b b b c c d d';
-  grid-template-columns: repeat(12, minmax(0, 1fr));
-  grid-template-rows: auto auto;
-  align-items: center;
-  padding: 18px 0;
-}
-
-.layout::before {
-  content: '';
-  position: absolute;
-  pointer-events: none;
-  inset: -2px -16px;
-  opacity: 0;
-  border-radius: 12px;
-  z-index: -1;
-}
-
-.layout:hover::before {
-  opacity: 1;
-}
-
-.area-a {
-  grid-area: a;
-}
-.area-b {
-  grid-area: b;
-}
-.area-c {
-  grid-area: c;
-}
-.area-d {
-  grid-area: d;
-  justify-self: end;
-}
-
-@media (max-width: 768px) {
-  .layout {
-    grid-template-columns: repeat(8, minmax(0, 1fr));
-    row-gap: 8px;
-    align-items: start;
-    grid-template-areas:
-      'a a a a a a c c'
-      'b b b b d d d d';
-  }
-
-  .area-c {
-    justify-self: end;
-  }
-}
